+++ src/lib/pages/execute/components/ExecuteArea.tsx
@@ -10,6 +10,7 @@ import {
   MessageTabs,
   UploadSchemaSection,
 } from "lib/components/json-schema";
+import { Tooltip } from "lib/components/Tooltip";
 import { useSchemaStore } from "lib/providers/store";
 import type { ContractAddr } from "lib/types";
 
@@ -33,9 +34,9 @@ export const ExecuteArea = observer(
     codeId,
   }: ExecuteAreaProps) => {
     const [tab, setTab] = useState<MessageTabs>();
-    const { getExecuteSchema } = useSchemaStore();
+    const { getExecuteSchema, getSchemaByCodeHash } = useSchemaStore();
     const schema = getExecuteSchema(codeHash);
-
+    const attached = Boolean(getSchemaByCodeHash(codeHash));
     const currentTabIdx = tab ? Object.values(MessageTabs).indexOf(tab) : 0;
 
     useEffect(() => {
@@ -58,7 +59,12 @@ export const ExecuteArea = observer(
               onClick={() => setTab(MessageTabs.YOUR_SCHEMA)}
               isDisabled={!contractAddress}
             >
-              Your Schema
+              <Tooltip
+                label="Please select contract first"
+                isDisabled={Boolean(contractAddress)}
+              >
+                Your Schema
+              </Tooltip>
             </CustomTab>
           </TabList>
         </Tabs>
@@ -72,8 +78,10 @@ export const ExecuteArea = observer(
             />
           }
           schemaContent={
-            codeHash && schema ? (
+            codeHash && attached ? (
               <SchemaExecute
+                codeId={codeId}
+                codeHash={codeHash}
                 schema={schema}
                 contractAddress={contractAddress}
                 initialFunds={initialFunds}
@@ -88,7 +96,7 @@ export const ExecuteArea = observer(
                     <Flex display="inline" textAlign="center">
                       You haven&#39;t attached the JSON Schema for
                       <CustomIcon name="code" mx={1} color="gray.400" />
-                      code {codeId} yet
+                      code {codeId}
                     </Flex>
                     <Flex textAlign="center">
                       from which this contract is instantiated yet.