+++ src/lib/pages/query/components/QueryArea.tsx
@@ -10,6 +10,7 @@ import {
   MessageTabs,
   UploadSchemaSection,
 } from "lib/components/json-schema";
+import { Tooltip } from "lib/components/Tooltip";
 import { useSchemaStore } from "lib/providers/store";
 import type { ContractAddr } from "lib/types";
 
@@ -27,8 +28,9 @@ export const QueryArea = observer(
   ({ contractAddress, codeId, codeHash, initialMsg }: QueryAreaProps) => {
     const [tab, setTab] = useState<MessageTabs>();
 
-    const { getQuerySchema } = useSchemaStore();
+    const { getQuerySchema, getSchemaByCodeHash } = useSchemaStore();
     const schema = getQuerySchema(codeHash);
+    const attached = Boolean(getSchemaByCodeHash(codeHash));
     const isMobile = useMobile();
 
     const currentTabIdx = tab ? Object.values(MessageTabs).indexOf(tab) : 0;
@@ -54,7 +56,12 @@ export const QueryArea = observer(
                 onClick={() => setTab(MessageTabs.YOUR_SCHEMA)}
                 isDisabled={!contractAddress}
               >
-                Your Schema
+                <Tooltip
+                  label="Please select contract first"
+                  isDisabled={Boolean(contractAddress)}
+                >
+                  Your Schema
+                </Tooltip>
               </CustomTab>
             </TabList>
           </Tabs>
@@ -68,8 +75,10 @@ export const QueryArea = observer(
             />
           }
           schemaContent={
-            codeHash && schema ? (
+            codeHash && attached ? (
               <SchemaQuery
+                codeId={codeId}
+                codeHash={codeHash}
                 schema={schema}
                 contractAddress={contractAddress}
                 initialMsg={initialMsg}
@@ -83,7 +92,7 @@ export const QueryArea = observer(
                     <Flex display="inline" textAlign="center">
                       You haven&#39;t attached the JSON Schema for
                       <CustomIcon name="code" mx={1} color="gray.400" />
-                      code {codeId} yet
+                      code {codeId}
                     </Flex>
                     <Flex textAlign="center">
                       from which this contract is instantiated yet.